requirements.txt
# 🎯 Learning Emotion Radar - Ultra Dependencies
# Real-time vision AI for educational emotion detection

# Core Streamlit and Web Framework
streamlit>=1.28.0

# IBM Watsonx AI - Vision Model Integration
ibm-watsonx-ai>=1.1.14

# Environment and Configuration
python-dotenv>=1.0.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Visualization and Charts
plotly>=5.15.0
pandas>=2.0.0

# Scientific Computing
numpy>=1.24.0

# Date and Time Utilities (usually built-in, but ensuring compatibility)
typing-extensions>=4.7.0

# Optional: Enhanced UI Components
streamlit-camera-input-live>=0.2.0  # For potential real-time webcam streaming

# Development and Testing (optional)
# pytest>=7.4.0
# jupyter>=1.0.0

# Production Deployment (optional)
# gunicorn>=21.2.0